// src/generated/github-actions.d.ts
export type StringContainingExpressionSyntax = string;
// ...
export type ExpressionSyntax = string;
// ...
export type Machine = 'linux' | 'macos' | 'windows';
export type Architecture = 'ARM32' | 'x64' | 'x86';
// ...
export interface NormalJob {
  /**
   * The name of the job displayed on GitHub.
   */
  name?: string;
  needs?: JobNeeds;
  permissions?: Permissions;
  /**
   * The type of machine to run the job on. The machine can be either a GitHub-hosted runner, or a self-hosted runner.
   */
  'runs-on':
    | (
        | 'macos-10.15'
        | 'macos-11'
        | 'macos-12'
        | 'macos-latest'
        | 'self-hosted'
        | 'ubuntu-18.04'
        | 'ubuntu-20.04'
        | 'ubuntu-22.04'
        | 'ubuntu-latest'
        | 'ubuntu-latest-4-cores'
        | 'ubuntu-latest-8-cores'
        | 'ubuntu-latest-16-cores'
        | 'windows-2019'
        | 'windows-2022'
        | 'windows-latest'
        | 'windows-latest-8-cores'
      )
    | ((
        | ['self-hosted', ...string[]]
        | ['self-hosted', Machine, ...string[]]
        | ['self-hosted', Architecture, ...string[]]
        | ['self-hosted', Machine, Architecture, ...string[]]
        | ['self-hosted', Architecture, Machine, ...string[]]
      ) &
        unknown[])
    | {
        group?: string;
        labels?: string | string[];
      }
    | StringContainingExpressionSyntax;
  /**
   * The environment that the job references.
   */
  environment?: string | Environment;
  /**
   * A map of outputs for a job. Job outputs are available to all downstream jobs that depend on this job.
   */
  outputs?: {
    [k: string]: string;
  };
// ...
    ) & {
// ...
      'continue-on-error'?: boolean | ExpressionSyntax;
      /**
       * The maximum number of minutes to run the step before killing the process.
       */
      'timeout-minutes'?: number | ExpressionSyntax;
    },
// ...
}